added "close" command
added initial setup

+++ bot.js
@@ -1,7 +1,10 @@
+'use strict';
 const Discord = require('discord.js');
 const client = new Discord.Client();
 const sqlite3 = require('sqlite3');
 const open = require('sqlite').open;
+const botKey = Math.floor(Math.random() * 100000000);
+let botChannel;
 let rows;
 let db;
 
@@ -17,6 +20,40 @@ let db;
 	                             where key = 'token';`));
 	if (token) {
 		await client.login(token['value']);
+		rows = await db.get(`select value
+		                     from general
+		                     where key = 'bot_channel'`);
+		if (rows) {
+			botChannel = rows['value'];
+			const channel = await client.channels.fetch(botChannel);
+			if (channel) {
+				rows = await db.get(`select value
+				                     from general
+				                     where key = 'first_startup'`);
+				if (rows ? rows['value'] !== "false" : true) {
+					console.log('First startup, bot channel found.');
+					channel.send('Hi! Call me roomBot! I will live here from now on!\n' +
+						'Talk to me by typing ">help" here!');
+					db.run(`insert into general
+					        values ('first_startup', 'false')`);
+				} else {
+					console.log('Bot channel found.');
+				}
+			} else {
+				console.log('Bot management channel deleted or unavailable. Use command ">channel ' +
+					botKey + '" (without quotes) in the desired channel to update.');
+				db.run(`delete
+				        from general
+				        where key = 'first_startup'`);
+			}
+		} else {
+			botChannel = "0";
+			console.log('Bot management channel not set. Use command ">channel '
+				+ botKey + '" (without quotes) in the desired channel to set.');
+			db.run(`delete
+			        from general
+			        where key = 'first_startup'`);
+		}
 	} else {
 		console.log('NO TOKEN FOUND!');
 	}
@@ -31,20 +68,20 @@ async function delay(ms) {
 
 
 client.on('ready', () => {
-	console.log('Ready');
+	console.log('Ready.');
 });
 
 
 client.on('message', async message => {
 	// Necessary checks for the bot to stay in its channel and answer only messages starting with '/'
-	if (message.content.startsWith('/')) {
+	if (message.content.startsWith('>')) {
 		const content = message.content;
 		const channel = message.channel;
 		const guild = channel.guild;
 		const channels = guild.channels;
 		
 		//await channel.startTyping();
-		if (channel.name === 'room-bot') {
+		if (channel.id === botChannel) {
 			// Room commands
 			if (content.includes('room')) {
 				const args = content.split(' ');
@@ -88,7 +125,7 @@ client.on('message', async message => {
 					channel.send(`There you go! Created room ${id} just for you!  :relaxed:`);
 				}
 				// Add or remove users from the room
-				else if (args[2] === 'add' || args[2] === 'kick') {
+				else if (args[2] === 'invite' || args[2] === 'kick') {
 					const roomId = args[1];
 					const username = args[3];
 					
@@ -103,7 +140,7 @@ client.on('message', async message => {
 					if (!room) {
 						channel.send('You sure the room exists at all? Can\'t find it...  ');
 					} else if (!user) {
-						channel.send(`You want to chat with ${username}? Can\'t find any ${username} in here  :worried:`)
+						channel.send(`Who are you talking about? I can\'t find any ${username} in here  :worried:`)
 					} else {
 						rows = await db.get(`select owner_snowflake as osf
 						                     from rooms
@@ -113,49 +150,66 @@ client.on('message', async message => {
 							channel.send('Someone is trying to mess with your room!!!  :sos: :sos: :sos:', {
 								reply: owner.user
 							});
-						}
-						
-						rows = await db.get(`select owner_snowflake, room_users.user_snowflake
-						                     from rooms
-							                          left join room_users on rooms.id = room_users.room_id
-						                     where channel_snowflake = $cid
-							                   and user_snowflake = $uid`, {$cid: room.id, $uid: user.user.id});
-						if (args[2] === 'add') {  // Adding user
-							
-							if (user.user.id === message.author.id) {
-								channel.send('Have you decided to add yourself again? Don\'t worry, ' +
-									'you\'re already there!  :wink:')
-							} else if (rows) {
-								channel.send('You\'re trying to add the same person again. Just go to the channel, ' +
-									'they\'re waiting for you there!');
-							} else {
-								db.run(`insert into room_users(room_id, user_snowflake)
-								        values ($rid, $uid)`, {$rid: roomId, $uid: user.user.id});
-								await room.updateOverwrite(user, {'VIEW_CHANNEL': true});
-								channel.send('Done, enjoy their company!  :fireworks:');
-							}
 						} else {
-							if (user.user.id === message.author.id) {
-								channel.send('Wait, if you remove yourself, then who\'s left to manage your room?  ' +
-									':scream:\nDon\'t even think about that! We need you!')
-							} else if (!rows) {
-								channel.send('You\'re trying to kick a person who\'s even not in the room! ' +
-									'Why do you hate them so much!?  :frowning:')
+							
+							rows = await db.get(`select owner_snowflake, room_users.user_snowflake
+							                     from rooms
+								                          left join room_users on rooms.id = room_users.room_id
+							                     where channel_snowflake = $cid
+								                   and user_snowflake = $uid`, {$cid: room.id, $uid: user.user.id});
+							if (args[2] === 'invite') {  // Adding user
+								
+								if (user.user.id === message.author.id) {
+									channel.send('Have you decided to add yourself again? Don\'t worry, ' +
+										'you\'re already there!  :wink:')
+								} else if (rows) {
+									channel.send('You\'re trying to add the same person again. Just go to the channel, ' +
+										'they\'re waiting for you there!  :partying_face:');
+								} else {
+									db.run(`insert into room_users(room_id, user_snowflake)
+									        values ($rid, $uid)`, {$rid: roomId, $uid: user.user.id});
+									await room.updateOverwrite(user, {'VIEW_CHANNEL': true});
+									channel.send(`Done, enjoy ${user.user.username}\'s company!  :fireworks:`);
+								}
 							} else {
-								db.run(`delete
-								        from room_users
-								        where room_id = $rid
-									      and user_snowflake = $uid`, {$rid: roomId, $uid: user.user.id});
-								await room.permissionOverwrites.get(user.user.id).delete();
-								channel.send('Done! They won\'t bother you ever again!  :smiling_imp:  ' +
-									'(well, at least, in that room)');
+								if (user.user.id === message.author.id) {
+									channel.send('Wait, if you remove yourself, then who\'s left to manage your room?  ' +
+										':scream:\nDon\'t even think about that! We need you!')
+								} else if (!rows) {
+									channel.send('You\'re trying to kick a person who\'s even not in the room! ' +
+										'Why do you hate them so much!?  :frowning:')
+								} else {
+									db.run(`delete
+									        from room_users
+									        where room_id = $rid
+										      and user_snowflake = $uid`, {$rid: roomId, $uid: user.user.id});
+									await room.permissionOverwrites.get(user.user.id).delete();
+									channel.send('Done! They won\'t bother you ever again!  :smiling_imp:  ' +
+										'(well, at least, in that room)');
+								}
 							}
 						}
 					}
 				}
 				// Remove the room
 				else if (args[2] === 'close') {
-					channel.send('I can\'t do this just yet.  :cry:  But I will, I promise!');
+					const roomId = args[1];
+					rows = await db.get(`select channel_snowflake as csf, owner_snowflake as osf
+					                     from rooms
+					                     where id = $rid`, {$rid: roomId});
+					if (!rows) {
+						channel.send('Let\' pretend I deleted your invisible room!  :upside_down:');
+					} else if (rows['osf'] !== message.author.id) {
+						channel.send('Trying to delete the room you didn\'t create! How naive! ' +
+							'Did you really think I\'ll let you do that!?  :face_with_monocle:')
+					} else {
+						db.run(`delete
+						        from rooms
+						        where id = $rid`, {$rid: roomId});
+						const room = (await guild.channels.resolve(rows['csf'])).delete();
+						channel.send('Let\'s forget these dark times and I promise that nobody will ' +
+							'ever see that again  :zipper_mouth:');
+					}
 				}
 				
 				// Commands not recognized by the bot
@@ -166,10 +220,45 @@ client.on('message', async message => {
 			
 			// Clear command
 			else if (content.includes('clear')) {
-				await channel.bulkDelete(100);
-				channel.send(`Enjoy your clear channel now, ${message.author.username}!  :relieved:`);
+				rows = await db.get(`select value
+				                     from general
+				                     where key = 'allow_clearing'`);
+				if (rows ? rows['value'] === 'true' : false) {
+					await channel.bulkDelete(100);
+					channel.send(`Enjoy your clear channel now, ${message.author.username}!  :relieved:`);
+				} else {
+					channel.send('I\'m not allowed to delete messages here. Some bad people want ' +
+						'everything to stay here forever...  :sweat:')
+				}
 			}
 			
+			// Help command
+			else if (content.includes('help')) {
+				channel.send('I am the roomBot and I can help you to create and manage private rooms on this ' +
+					'server!  :slight_smile:\n' +
+					'  1. To open the new room, type ">room open"\n' +
+					'  2. To invite people to your room, type ">room [room number] invite [user]"\n' +
+					'  3. To kick someone from your room, type ">room [room number] kick [user]"\n' +
+					'  4. To close the room, type ">room [room number] close"\n' +
+					'  5. On some servers you can use ">clear" command to clear *this* channel.' +
+					'That\'s it! Enjoy!  :star_struck:');
+			}
+		}
+		
+		// Channel command
+		else if (content.includes('channel')) {
+			const args = content.split(' ');
+			if (args[1] === botKey.toString()) {
+				db.run(`insert into general
+				        values ($key, $value)`, {$key: 'bot_channel', $value: channel.id});
+				console.log('Bot channel successfully set. Restart the bot now.');
+				channel.send('Channel set. Restart the bot now.');
+			} else {
+				console.log('WARNING: somebody tried to change bot channel but used the wrong key!');
+				channel.send('Wrong key!');
+			}
+			await delay(5000);
+			await channel.bulkDelete(2);
 		} else {
 			channel.send('Were you looking for me? Ugh, I have no powers here...  :sob:  Meet me in my channel!')
 		}
